Remove loader-utils dependency
This was the only dependency of this package. It provides several helper
functions. This package was using `interpolateName` and `getOptions`.

The first one is not really needed since it only allows to configure a
name using placeholders (like `[name].[ext]`).

The second one allows to merge options with default values which can
be achieved manually using `this.query` (from webpack). It won't work
on the non-hashes but this is simple enough to debug (and README.md
includes how the custom options should look like).